Overhaul dotplot highlighting
A much more complex approach to code (using a custom method forteh
ggplot "+" operator), which provides a much simpler API to the user.

From this commit forward, `highlight_query` and `highlight_target`
inherit chromsome-ordering and alignment data from the parent dotplot
(these argemetn rpeviously had to be provided and match the same values
used in the original plot).

Also updates docs, vignette and tests to account for change in API